WEB-2039 start moving files
