+++ src/sheet-common.tsx
@@ -1,7 +1,7 @@
 'use client';
 import classnames from 'classnames';
 import * as React from 'react';
-import * as styles from './sheet.css';
+import * as styles from './sheet-common.css';
 import FocusTrap from './focus-trap';
 import {useDisableBodyScroll, useIsInViewport, useScreenSize, useTheme} from './hooks';
 import {useSetModalStateEffect} from './modal-context-provider';
@@ -14,8 +14,6 @@ import NegativeBox from './negative-box';
 import Stack from './stack';
 import Box from './box';
 import Touchable from './touchable';
-import Inline from './inline';
-import Circle from './circle';
 import Divider from './divider';
 import {getPrefixedDataAttributes, getScrollableParentElement} from './utils/dom';
 import {ButtonLink, ButtonPrimary, ButtonSecondary} from './button';
@@ -575,90 +573,6 @@ export const ActionsListSheet = React.forwardRef<HTMLDivElement, ActionsListShee
                                 </Touchable>
                             ))}
                         </NegativeBox>
-                    </SheetBody>
-                )}
-            </Sheet>
-        );
-    }
-);
-
-type InfoSheetProps = {
-    title?: string;
-    subtitle?: string;
-    description?: string | ReadonlyArray<string>;
-    items: ReadonlyArray<{
-        id?: string;
-        title: string;
-        description?: string;
-        icon: ExclusifyUnion<
-            | {
-                  type: 'regular' | 'small';
-                  Icon: React.ComponentType<IconProps>;
-              }
-            | {
-                  type: 'regular' | 'small';
-                  url: string;
-                  urlDark?: string;
-              }
-            | {type: 'bullet'}
-        >;
-    }>;
-    onClose?: () => void;
-    dataAttributes?: DataAttributes;
-};
-
-export const InfoSheet = React.forwardRef<HTMLDivElement, InfoSheetProps>(
-    ({title, subtitle, description, items, onClose, dataAttributes}, ref) => {
-        const {isDarkMode} = useTheme();
-        return (
-            <Sheet
-                onClose={onClose}
-                ref={ref}
-                dataAttributes={{'component-name': 'InfoSheet', ...dataAttributes}}
-            >
-                {({modalTitleId}) => (
-                    <SheetBody
-                        title={title}
-                        subtitle={subtitle}
-                        description={description}
-                        modalTitleId={modalTitleId}
-                    >
-                        <Box paddingBottom={16}>
-                            <Stack space={16} role="list">
-                                {items.map((item, idx) => (
-                                    <Inline key={item.id || idx} space={8}>
-                                        <div className={styles.infoItemIcon}>
-                                            {item.icon.type === 'bullet' ? (
-                                                <Circle
-                                                    size={8}
-                                                    backgroundColor={skinVars.colors.textPrimary}
-                                                />
-                                            ) : item.icon.Icon ? (
-                                                <item.icon.Icon size={item.icon.type === 'small' ? 16 : 24} />
-                                            ) : (
-                                                <Image
-                                                    src={
-                                                        isDarkMode && item.icon.urlDark
-                                                            ? item.icon.urlDark
-                                                            : item.icon.url
-                                                    }
-                                                    width={item.icon.type === 'small' ? 16 : 24}
-                                                    height={item.icon.type === 'small' ? 16 : 24}
-                                                />
-                                            )}
-                                        </div>
-                                        <Stack space={2}>
-                                            <Text3 regular>{item.title}</Text3>
-                                            {item.description && (
-                                                <Text2 regular color={skinVars.colors.textSecondary}>
-                                                    {item.description}
-                                                </Text2>
-                                            )}
-                                        </Stack>
-                                    </Inline>
-                                ))}
-                            </Stack>
-                        </Box>
                     </SheetBody>
                 )}
             </Sheet>
